requirements.txt
numpy==1.13.3
Pillow==6.0.0
torch==1.1.0
torchvision==0.2.2.post3
